Moved the DataProviderType tests into test_dataprovider.py
Also test that the returned value in the enum is the expected class
